requirements.txt
Pytorch
pandas
numpy